created seqtools utilities from proteins.compare and sequence
